requirements.txt
amqp==2.1.3
argparse==1.2.1
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
celery==4.0.1
cffi==1.9.1
click==6.6
cryptography==1.7.1
enum34==1.1.6
Flask==0.11.1
Flask-SQLAlchemy==2.1
idna==2.2
ipaddress==1.0.17
itsdangerous==0.24
Jinja2==2.8
kombu==4.0.1
MarkupSafe==0.23
ndg-httpsclient==0.4.2
nose==1.3.7
pyaml==16.9.0
pyasn1==0.1.9
pycparser==2.17
pymongo==3.4.0
pyOpenSSL==16.2.0
python-dateutil==2.6.0
pytz==2016.10
PyYAML==3.12
redis==2.10.5
requests==2.11.1
simplejson==3.10.0
six==1.10.0
slackclient==1.0.2
SQLAlchemy==1.1.4
vine==1.1.3
websocket-client==0.37.0
Werkzeug==0.11.11
wit==4.2.0
wsgiref==0.1.2